[Commons] Game Of Life Code Challenge
This is the first commit of the GoF game challenge.

Candidates are asked to build comway's game of life in Ruby as part of
the interview process.

Here is specifically what we're looking for (taken from the "problem solving",
"code fluency" and "system design" sections of our on site interviewing rubric):

  - A working solution covering many edge cases.
  - Well named components (classes, methods, variables etc). We care about
    deliberate and intentional naming of components.
  - Idiomatic Ruby.
  - Well separated concerns with good interfaces.
  - Intuitive understanding of SOLID principles.
  - Elegant components that are well separated and reusable.
  - Clear abstractions that facilitate extensions in the future without significant
    refactoring.
  - Details like idempotency and invariants considered if relevant to the problem.