[GoL] Adds Grid Class
This class represents the universe as describe for the game of life.
The Grid will manage the placement of cells.

- Adds support for building a Grid based on a Matrix

Decided to subclass a Matrix to borrow some useful functionality, for
instance the Matrix while immutable will allow for easy Cell, row and
column access.

This is a starting point and it might be possible we have to deviate as
we learn more about how the Game Of Life behaves.