[GoL] Gets Living Neighbors Count By Cell
The grid now can find a count of neighboring living cells for a given
cell.

Adds a Neighbor module:
This module performs the calculations to determine the neighborhood.

I decided to separate this concern so it is easier to implementent
or use another algorithm to search for neighbors in case the logic
changes in the future, being aware that the Game Of Life has strict
rules to find neighbors.