[GoL] Adds Live Cell Rule
As per GoL there are some rules that needs to be applied for each cell
in order to determine if the cell will survive to the next generation.

This adds the Live rules of the game which covers

* Any live cell with fewer than two live neighbors dies, as if caused by under
  population. (cell < 2 live neighbors = dies)
* Any live cell with two or three live neighbors lives on to the next generation.
  (cell with 2,3 live neighbors = survives)
* Any live cell with more than three live neighbors dies, as if by overpopulation.
  (cell > 3 live neighbors = dies)

In the following PRs we will be implementing some form of the State
Design Pattern to determine which rule to apply.