[GoL] The game of life is ready! Yaay!
This was a fun code challenge and I learned a lot!

I did a lot of reasearch before and analysis to implement it the best
way possible. There is a few gotchas in this implementation that maybe
you would like to discuss, but overall I feel confident with the apprach
even though there is room for improvements, specially in the rendering
time of the animation, see the notes section below for more details.

Changes:

- A `bin/run` file so that you can test it locally in your console.
- Adds the `tick` method to the grid
- Adds a play method to the GameOfLife Module

Notes/Side Effects:

I notice that the demos seems to be very slow and one of the reasons
for this has to do with the fact that we have to loop over the `@cells`
variable to determine if the cell is alive, to improve this it would be
great to determine the state in O(1) time. A way to do this is
saving state in the cell or instead of an array to use a Hash.